crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.1",
    "langchain>=1.2.12",
    "langgraph>=1.1.2",
    "langgraph-api[inmem]>=0.7.71",
    "langgraph-cli[inmem]>=0.4.18",
    "replicate>=0.30.0",
    "httpx>=0.27.0",
    "crawl4ai>=0.3.7",
    "motor>=3.4.0",
    "neo4j>=5.19.0",
    "chromadb>=0.5.0",
    "pydantic>=2.7.0",
    "langchain-core>=0.1.0",
    "langchain-tavily>=0.1.0",
    "mcp>=1.26.0",
    "google-search-results>=2.4.2",
    "langchain-openai>=1.1.11",
    "streamlit>=1.55.0",
    "pandas>=2.3.3",
]
